requirements.txt
Django==1.9
django-cors-headers==2.1.0
django-filter==1.1.0
djangorestframework==3.7.3
Markdown==2.6.10
pytz==2017.3
